technical_demo/requirements.txt
web3~=5.17.0
py-solc-x~=1.1.0

networkx~=2.5
matplotlib~=3.3.4